main - use mainRunner and existing cleanup
